Updated isend and send

+++ src/receive.rs
@@ -6,13 +6,13 @@ use std::sync::Arc;
 
 // Functions in the Receive module
 pub fn mpi_irecv<'a, T: AsRef<&'a [u8]>>(count: u64, datatype: MPIDatatype,
-			 dest: u64, tag: u64, comm: MPIComm, request: &mut MPIRequest) -> Arc<T> {
+			 dest: usize, tag: u64, comm: MPIComm, request: &mut MPIRequest) -> Arc<T> {
 
 	unimplemented!();
 	// let data = base64::from_base64();
 }
 
 pub fn mpi_recv<'a, T: AsRef<&'a [u8]>>(count: u64, datatype: MPIDatatype,
-			 dest: u64, tag: u64, comm: MPIComm) -> Arc<T> {
+			 dest: usize, tag: u64, comm: MPIComm) -> Arc<T> {
 	unimplemented!();
 }
